+++ server/impl/src/main/java/com/walmartlabs/concord/server/A.java
@@ -1,4 +1,0 @@
-package com.walmartlabs.concord.server;
-
-public class A {
-}